+++ src/config/setup/config_setup_autoscan.cc
@@ -67,24 +67,33 @@ std::string ConfigAutoscanSetup::getItemPathRoot(bool prefix) const
 }
 
 /// @brief Creates an array of AutoscanDirectory objects from a XML nodeset.
-bool ConfigAutoscanSetup::createOptionFromNode(const pugi::xml_node& element, std::vector<std::shared_ptr<AutoscanDirectory>>& result)
+bool ConfigAutoscanSetup::createOptionFromNode(
+    const std::shared_ptr<Config>& config,
+    const pugi::xml_node& element,
+    std::vector<std::shared_ptr<AutoscanDirectory>>& result)
 {
     if (!element)
         return true;
 
     auto&& cs = definition->findConfigSetup<ConfigSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY);
+    if (config) {
+        config->registerNode(element.path());
+    }
     for (auto&& it : cs->getXmlTree(element)) {
         const pugi::xml_node& child = it.node();
+        if (config) {
+            config->registerNode(child.path());
+        }
 
         fs::path location;
         try {
-            location = definition->findConfigSetup<ConfigPathSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_LOCATION)->getXmlContent(child, true);
+            location = definition->findConfigSetup<ConfigPathSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_LOCATION)->getXmlContent(child, config, true);
         } catch (const std::runtime_error&) {
             log_warning("Found an Autoscan directory with invalid location!");
             continue;
         }
 
-        AutoscanScanMode mode = definition->findConfigSetup<ConfigEnumSetup<AutoscanScanMode>>(ConfigVal::A_AUTOSCAN_DIRECTORY_MODE)->getXmlContent(child);
+        AutoscanScanMode mode = definition->findConfigSetup<ConfigEnumSetup<AutoscanScanMode>>(ConfigVal::A_AUTOSCAN_DIRECTORY_MODE)->getXmlContent(child, config);
 
         if (mode != scanMode) {
             continue; // skip scan modes that we are not interested in (content manager needs one mode type per array)
@@ -92,25 +101,25 @@ bool ConfigAutoscanSetup::createOptionFromNode(const pugi::xml_node& element, st
 
         long long interval = 0;
         if (mode == AutoscanScanMode::Timed) {
-            interval = definition->findConfigSetup<ConfigTimeSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_INTERVAL)->getXmlContent(child);
+            interval = definition->findConfigSetup<ConfigTimeSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_INTERVAL)->getXmlContent(child, config);
         }
 
-        bool recursive = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_RECURSIVE)->getXmlContent(child);
-        bool forceRescan = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_FORCE_REREAD_UNKNOWN)->getXmlContent(child);
-        bool dirtypes = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_DIRTYPES)->getXmlContent(child);
-        int mt = definition->findConfigSetup<ConfigIntSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_MEDIATYPE)->getXmlContent(child);
+        bool recursive = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_RECURSIVE)->getXmlContent(child, config);
+        bool forceRescan = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_FORCE_REREAD_UNKNOWN)->getXmlContent(child, config);
+        bool dirtypes = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_DIRTYPES)->getXmlContent(child, config);
+        int mt = definition->findConfigSetup<ConfigIntSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_MEDIATYPE)->getXmlContent(child, config);
         log_debug("mt = {} -> {}", mt, AutoscanDirectory::mapMediaType(mt));
 
-        unsigned int retryCount = definition->findConfigSetup<ConfigUIntSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_RETRYCOUNT)->getXmlContent(child);
+        unsigned int retryCount = definition->findConfigSetup<ConfigUIntSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_RETRYCOUNT)->getXmlContent(child, config);
         auto cs = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_HIDDENFILES);
-        bool hidden = cs->hasXmlElement(child) ? cs->getXmlContent(child) : hiddenFiles;
+        bool hidden = cs->hasXmlElement(child) ? cs->getXmlContent(child, config) : hiddenFiles;
 
         cs = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_AUTOSCAN_DIRECTORY_FOLLOWSYMLINKS);
-        bool follow = cs->hasXmlElement(child) ? cs->getXmlContent(child) : followSymlinks;
+        bool follow = cs->hasXmlElement(child) ? cs->getXmlContent(child, config) : followSymlinks;
 
-        auto ctAudio = definition->findConfigSetup<ConfigStringSetup>(ConfigVal::A_AUTOSCAN_CONTAINER_TYPE_AUDIO)->getXmlContent(child);
-        auto ctImage = definition->findConfigSetup<ConfigStringSetup>(ConfigVal::A_AUTOSCAN_CONTAINER_TYPE_IMAGE)->getXmlContent(child);
-        auto ctVideo = definition->findConfigSetup<ConfigStringSetup>(ConfigVal::A_AUTOSCAN_CONTAINER_TYPE_VIDEO)->getXmlContent(child);
+        auto ctAudio = definition->findConfigSetup<ConfigStringSetup>(ConfigVal::A_AUTOSCAN_CONTAINER_TYPE_AUDIO)->getXmlContent(child, config);
+        auto ctImage = definition->findConfigSetup<ConfigStringSetup>(ConfigVal::A_AUTOSCAN_CONTAINER_TYPE_IMAGE)->getXmlContent(child, config);
+        auto ctVideo = definition->findConfigSetup<ConfigStringSetup>(ConfigVal::A_AUTOSCAN_CONTAINER_TYPE_VIDEO)->getXmlContent(child, config);
         try {
             auto containerMap = AutoscanDirectory::ContainerTypesDefaults;
             containerMap[AutoscanMediaMode::Audio] = ctAudio;
@@ -299,7 +308,10 @@ bool ConfigAutoscanSetup::updateDetail(const std::string& optItem, std::string& 
     return false;
 }
 
-void ConfigAutoscanSetup::makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments)
+void ConfigAutoscanSetup::makeOption(
+    const pugi::xml_node& root,
+    const std::shared_ptr<Config>& config,
+    const std::map<std::string, std::string>* arguments)
 {
     if (arguments && arguments->find("hiddenFiles") != arguments->end()) {
         hiddenFiles = arguments->at("hiddenFiles") == "true";
@@ -307,14 +319,16 @@ void ConfigAutoscanSetup::makeOption(const pugi::xml_node& root, const std::shar
     if (arguments && arguments->find("followSymlinks") != arguments->end()) {
         followSymlinks = arguments->at("followSymlinks") == "true";
     }
-    newOption(getXmlElement(root));
+    newOption(config, getXmlElement(root));
     setOption(config);
 }
 
-std::shared_ptr<ConfigOption> ConfigAutoscanSetup::newOption(const pugi::xml_node& optValue)
+std::shared_ptr<ConfigOption> ConfigAutoscanSetup::newOption(
+    const std::shared_ptr<Config>& config,
+    const pugi::xml_node& optValue)
 {
     auto result = std::vector<std::shared_ptr<AutoscanDirectory>>();
-    if (!createOptionFromNode(optValue, result)) {
+    if (!createOptionFromNode(config, optValue, result)) {
         throw_std_runtime_error("Init {} autoscan failed '{}'", xpath, optValue.name());
     }
     optionValue = std::make_shared<AutoscanListOption>(result);