+++ src/config/setup/config_setup_vector.cc
@@ -36,11 +36,12 @@
 
 /// @brief Creates a vector from an XML nodeset.
 bool ConfigVectorSetup::createOptionFromNode(
+    const std::shared_ptr<Config>& config,
     const pugi::xml_node& element,
     std::vector<std::vector<std::pair<std::string, std::string>>>& result)
 {
     if (element) {
-        doExtend = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_LIST_EXTEND)->getXmlContent(element);
+        doExtend = definition->findConfigSetup<ConfigBoolSetup>(ConfigVal::A_LIST_EXTEND)->getXmlContent(element, config);
         const auto dictNodes = element.select_nodes(definition->mapConfigOption(nodeOption));
         std::vector<std::string> attrList;
         attrList.reserve(optionList.size());
@@ -50,10 +51,16 @@ bool ConfigVectorSetup::createOptionFromNode(
 
         for (auto&& it : dictNodes) {
             const pugi::xml_node child = it.node();
+            if (config) {
+                config->registerNode(child.path());
+            }
             std::vector<std::pair<std::string, std::string>> valueList;
             for (auto& attr : child.attributes()) {
                 std::string name = attr.name();
                 std::string value = attr.value();
+                if (config) {
+                    config->registerNode(fmt::format("{}/attribute::{}", child.path(), name));
+                }
                 if (!value.empty()) {
                     if (tolower) {
                         toLowerInPlace(value);
@@ -85,7 +92,7 @@ void ConfigVectorSetup::makeOption(
     if (arguments && arguments->find("tolower") != arguments->end()) {
         tolower = arguments->at("tolower") == "true";
     }
-    newOption(getXmlContent(getXmlElement(root)));
+    newOption(getXmlContent(getXmlElement(root), config));
     setOption(config);
 }
 
@@ -203,10 +210,12 @@ std::string ConfigVectorSetup::getItemPathRoot(bool prefix) const
     return fmt::format("{}/{}", xpath, definition->mapConfigOption(nodeOption));
 }
 
-std::vector<std::vector<std::pair<std::string, std::string>>> ConfigVectorSetup::getXmlContent(const pugi::xml_node& optValue)
+std::vector<std::vector<std::pair<std::string, std::string>>> ConfigVectorSetup::getXmlContent(
+    const pugi::xml_node& optValue,
+    const std::shared_ptr<Config>& config)
 {
     std::vector<std::vector<std::pair<std::string, std::string>>> result;
-    if (!createOptionFromNode(optValue, result) && required) {
+    if (!createOptionFromNode(config, optValue, result) && required) {
         throw_std_runtime_error("Init {} vector failed '{}'", xpath, optValue.name());
     }
     if (result.empty()) {